+++ CHANGELOG.rst
@@ -6,7 +6,10 @@ This document describes changes between each past release.
 3.3.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+**Bug fixes**
+
+- Fix crash with PostgreSQL storage backend when provided id in POST is an integer (#688).
+  Regression introduced in 3.2.0 with #655.
 
 
 3.2.0 (2016-06-14)