+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Protocol is now at version **1.8**. See `API changelog <http://kinto.readthedocs
 
 **Bug fixes**
 
+- Fix crash in authorization policy when requesting ``GET /buckets/collections`` (fixes #695)
 - Fix crash with PostgreSQL storage backend when provided id in POST is an integer (#688).
   Regression introduced in 3.2.0 with #655.
 - Fix crash with PostgreSQL storage backend is configured as read-only and reaching