+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Changelog
 
 This document describes changes between each past release.
 
+3.2.1 (unreleased)
+==================
+
+- Nothing changed yet.
+
 3.2.0 (2016-06-14)
 ==================